Get SoftwareId Directly in Sp Fixed it.

+++ Adroit.Accounting/Adroit.Accounting.Repository/IRepository/IProductAmtCalcOn.cs
@@ -10,6 +10,6 @@ namespace Adroit.Accounting.Repository.IRepository
         void Delete(int id, string connectionString);
         int Save(ProductAmtCalcOn value, string connectionString);
         ProductAmtCalcOnViewModel Get(int id, string connectionString);
-        List<DropdownViewModel> SelectList(byte softwareId, string connectionString);
+        List<DropdownViewModel> SelectList(int softwareId, string connectionString);
     }
 }
