+++ Adroit.Accounting/Adroit.Accounting.Repository/ProductAmtCalcOnRepository.cs
@@ -46,7 +46,7 @@ namespace Adroit.Accounting.Repository
             return QueryHelper.GetList<ProductAmtCalcOnGridViewModel>("sp_ProductAmtCalcOnList", connectionString, param);
         }
 
-        public List<DropdownViewModel> SelectList(byte softwareId, string connectionString)
+        public List<DropdownViewModel> SelectList(int softwareId, string connectionString)
         {
             var parameters = new DynamicParameters();
             parameters.Add("@softwareId", softwareId);
